[#3168] - add functionality for searching for datapoints
